adding testing to form login
